Improve augementation in ckeditor5-style.
